+++ Routinus/Routinus/Data/ChallengeRepository.swift
@@ -31,12 +31,6 @@ protocol ChallengeRepository {
               authExampleImageURL: String,
               authExampleThumbnailImageURL: String)
     func update(challenge: Challenge)
-    func updateImage(challengeID: String,
-                     imageURL: String,
-                     thumbnailImageURL: String)
-    func updateImage(challengeID: String,
-                     authExampleImageURL: String,
-                     authExampleThumbnailImageURL: String)
 }
 
 extension RoutinusRepository: ChallengeRepository {
@@ -154,33 +148,5 @@ extension RoutinusRepository: ChallengeRepository {
                                         ownerID: challenge.ownerID)
         RoutinusNetwork.updateChallenge(challengeDTO: challengeDTO,
                                         completion: nil)
-    }
-
-    func updateImage(challengeID: String,
-                     imageURL: String,
-                     thumbnailImageURL: String) {
-        RoutinusNetwork.uploadImage(id: challengeID,
-                                    filename: "image",
-                                    imageURL: imageURL,
-                                    completion: nil)
-        RoutinusNetwork.uploadImage(id: challengeID,
-                                    filename: "thumbnail_image",
-                                    imageURL: thumbnailImageURL,
-                                    completion: nil)
-        RoutinusStorage.removeCachedImages(from: challengeID)
-    }
-
-    func updateImage(challengeID: String,
-                     authExampleImageURL: String,
-                     authExampleThumbnailImageURL: String) {
-        RoutinusNetwork.uploadImage(id: challengeID,
-                                    filename: "auth",
-                                    imageURL: authExampleImageURL,
-                                    completion: nil)
-        RoutinusNetwork.uploadImage(id: challengeID,
-                                    filename: "thumbnail_auth",
-                                    imageURL: authExampleThumbnailImageURL,
-                                    completion: nil)
-        RoutinusStorage.removeCachedImages(from: challengeID)
     }
 }
